Consolidate tests a bit to reduce the amount of code in the PR
